The purpose of this commit is to address several review statements from PR #321     1. This commit changes the Visualizer Class names from ClassBalanceHeatMap to ClassPredictionError     2. ClassPredictError now resides as it's own visualizer in the class_balance.py module     3. I've affixed estimator attributes, such as self.classes_, with an underscore suffix     4. I've Used _check_target func to check if y_true and y_pred belong to the same classification task
