+++ yellowbrick/classifier/__init__.py
@@ -23,11 +23,10 @@ Visualizations related to evaluating Scikit-Learn classification models.
 ## Hoist visualizers into the classifier namespace
 from ..base import ScoreVisualizer
 from .base import ClassificationScoreVisualizer
-from .class_balance import ClassBalance
+from .class_balance import ClassBalance, ClassPredictionError
 from .classification_report import ClassificationReport, classification_report
 from .confusion_matrix import ConfusionMatrix
 from .learning_curve import LearningCurveVisualizer, learning_curve_plot
 from .rocauc import ROCAUC, roc_auc
 from .boundaries import decisionviz, DecisionBoundariesVisualizer, DecisionViz
 from .threshold import ThreshViz, ThresholdVisualizer, thresholdviz
-from .class_balance_heat_map import ClassBalanceHeatMap, class_balance_heat_map
